+++ src/test/java/com/ibanfr/shoppingbasket/ItemTest.java
@@ -3,6 +3,8 @@ package com.ibanfr.shoppingbasket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ItemTest {
@@ -12,7 +14,7 @@ class ItemTest {
     void should_create_an_Item_with_its_unit_price() {
 
         //given
-        int unitPrice = 100;
+        BigDecimal unitPrice = BigDecimal.valueOf(100);
 
         //when
         Item item = new Item(unitPrice);
